通讯连通，可进行拆包处理

+++ Source/FoolishGames.Common/Interfaces/IO/IMessageHeader.cs
@@ -12,7 +12,7 @@ namespace FoolishGames.IO
         /// <summary>
         /// 消息Id
         /// </summary>
-        int MsgId { get; }
+        long MsgId { get; }
 
         /// <summary>
         /// 操作码
